requirements.txt
# pip install -r requirements.txt
absl-py==0.11.0
astor==0.8.1
bleach==1.5.0
cached-property==1.5.2
certifi==2020.6.20
cycler==0.10.0
enum34==1.1.10
gast==0.4.0
grpcio==1.33.2
h5py==3.1.0
html5lib==0.9999999
importlib-metadata==2.0.0
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.16.0
Pillow==8.0.1
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.2.1
six==1.15.0
opencv-python==3.4.1.15
tensorboard==1.7.0
tensorflow-cpu==1.4.0
tensorflow-tensorboard==0.4.0
streamlit==1.10.0
termcolor==1.1.0
Werkzeug==1.0.1
zipp==3.4.0
